requirements.txt
quart==0.17.0
websockets==11.0.3
python-dotenv==1.0.0